Better module/type organization.
