+++ assets/src/components/v2/elevator/current_elevator_closed.tsx
@@ -5,7 +5,7 @@ import makePersistent, {
   WrappedComponentProps,
 } from "Components/v2/persistent_wrapper";
 import PagingIndicators from "Components/v2/elevator/paging_indicators";
-import { type Closure } from "Components/v2/elevator/models/closure";
+import { type Closure } from "Components/v2/elevator/types";
 import useClientPaging from "Hooks/v2/use_client_paging";
 import useTextResizer from "Hooks/v2/use_text_resizer";
 import CurrentLocationMarker from "Images/svgr_bundled/current-location-marker.svg";
