--- assets/src/components/v2/elevator/models/closure.ts
@@ -1,7 +1,0 @@
-export type Closure = {
-  id: string;
-  elevator_name: string;
-  elevator_id: string;
-  description: string;
-  header_text: string;
-};
