+++ assets/src/components/v2/elevator/types.ts
@@ -1,9 +1,16 @@
 import { type Pill } from "Components/v2/departures/route_pill";
-import { type Closure } from "Components/v2/elevator/models/closure";
 
 export type StationWithClosures = {
   id: string;
   name: string;
   route_icons: Pill[];
   closures: Closure[];
+};
+
+export type Closure = {
+  id: string;
+  elevator_name: string;
+  elevator_id: string;
+  description: string;
+  header_text: string;
 };
